Periodic checkin: coverted a fair bit of code into Kotlin.
I've been working through design.py and profile.py. Profile is, I believe,
fully converted; design is only about halfway done.  The original code is
pretty frighteningly impenetrable; I'm _mostly_ just slavishly copying
the original execution flow, rendering it into Kotlin and trying to apply
reasonable typing to it (which is not an easy feat.)

There are no tests yet, because honestly I don't understand what this
beast is doing well enough to be able to test it. As I continue to work
through the code, and through a couple of textbooks on acoustics, I'll
hopefully get to the point of understanding *enough* to write at least
a few reasonable tests.